pull up properties to common abstact type
